backend/pyproject.toml
[tool.poetry]
name = "apollo-commerce-api"
version = "0.1.0"
description = "Backend for the Apollo Commerce project"
authors = ["William Fernandes Dias <william.winchester1967@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.3"
django = "^4.2.6"
djangorestframework = "^3.14.0"
django-filter = "^23.3"
django-environ = "^0.11.2"
psycopg = {extras = ["binary"], version = "^3.1.12"}

[tool.poetry.group.code.dependencies]
mypy = "^1.6.1"
black = "^23.10.0"
django-stubs = { extras = ["compatible-mypy"], version = "^4.2.5" }
djangorestframework-stubs = { extras = [
    "compatible-mypy",
], version = "^3.14.4" }
isort = "^5.12.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"
pytest-mock = "^3.12.0"
hypothesis = { extras = ["django"], version = "^6.88.1" }
pytest-drf = "^1.1.3"

[build-system]
requires = ["poetry-core==1.5.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
skip_gitignore = true
